Fixed units in swigInputConfig.py for new Alchemical AMOEBA vdW methods exposed in the OpenMM API; implemented the getDefaultParameters() on the AmoebaVdwForceImpl to return 1.0 for AmoebaVdwAlpha; Use AmoebaVdwForce::Lambda() to refer to this parameter rather than the String AmoebaVdwAlpha

+++ plugins/amoeba/openmmapi/include/openmm/AmoebaVdwForce.h
@@ -258,12 +258,12 @@ public:
     /**
      * Set the softcore power on lambda (default = 5).
      */
-    void setSoftcorePower(double n);
+    void setSoftcorePower(int n);
 
     /**
      * Get the softcore power on lambda.
      */
-    double getSoftcorePower() const;
+    int getSoftcorePower() const;
 
     /**
      * Set the softcore alpha value (default = 0.7).
